requirements.txt
# Dev environment
pip
autopep8

# App
flask
pymongo 
google-generativeai 
python-dotenv 
twilio # Or equivalent for your chosen telephony service
ngrok